NoteByEtudiantId

+++ src/main/java/com/aclab/dne/repositories/NoteRepository.java
@@ -1,11 +1,14 @@
 package com.aclab.dne.repositories;
 
+import com.aclab.dne.dto.NoteDTO;
 import com.aclab.dne.model.Inscription;
+import com.aclab.dne.model.InscriptionId;
 import com.aclab.dne.model.Note;
 import org.springframework.data.repository.CrudRepository;
 
-public interface NoteRepository extends CrudRepository<Note,Integer> {
+import java.util.Optional;
+
+public interface NoteRepository extends CrudRepository<Note,Long> {
 
     Iterable<Note> findNotesByInscription(Inscription inscription);
-
 }
